+++ aoo/src/source.cpp
@@ -341,15 +341,15 @@ AooError AOO_CALL aoo::Source::control(
     // set/get time DLL filter bandwidth
     case kAooCtlSetDllBandwidth:
     {
-        CHECKARG(float);
-        auto bw = std::max<double>(0, std::min<double>(1, as<float>(ptr)));
+        CHECKARG(double);
+        auto bw = std::max<double>(0, std::min<double>(1, as<double>(ptr)));
         dll_bandwidth_.store(bw);
         reset_timer();
         break;
     }
     case kAooCtlGetDllBandwidth:
-        CHECKARG(float);
-        as<float>(ptr) = dll_bandwidth_.load();
+        CHECKARG(double);
+        as<double>(ptr) = dll_bandwidth_.load();
         break;
     case kAooCtlResetDll:
         reset_timer();
